requirements.txt
dnspython>=2.3.0
fastapi>=0.94.1
Jinja2>=3.1.2
uvicorn>=0.21.1
